Compilable rocsparse.
