+++ Manifest.toml
@@ -1,6 +1,6 @@
 # This file is machine-generated - editing it directly is not advised
 
-julia_version = "1.7.2"
+julia_version = "1.7.3"
 manifest_format = "2.0"
 project_hash = "bbe632dbaaf9e978eb1bd2b0b87a9ec9742fa931"
 
@@ -81,7 +81,7 @@ uuid = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
 version = "0.8.6"
 
 [[deps.Downloads]]
-deps = ["ArgTools", "LibCURL", "NetworkOptions"]
+deps = ["ArgTools", "FileWatching", "LibCURL", "NetworkOptions"]
 uuid = "f43a241f-c20a-4ad4-852c-f6b1247861c6"
 
 [[deps.Elfutils_jll]]
@@ -94,6 +94,9 @@ version = "0.182.0+1"
 git-tree-sha1 = "56559bbef6ca5ea0c0818fa5c90320398a6fbf8d"
 uuid = "e2ba6199-217a-4e67-a87a-7c52f15ade04"
 version = "0.1.8"
+
+[[deps.FileWatching]]
+uuid = "7b1f6079-737a-58dc-b8bc-7a2ca5c1b5ee"
 
 [[deps.Future]]
 deps = ["Random"]
@@ -296,9 +299,9 @@ version = "4.2.0+0"
 
 [[deps.ROCmDeviceLibs_jll]]
 deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg", "Zlib_jll"]
-git-tree-sha1 = "8f034123ed944497b8177e63963a034eeb4e62fa"
+git-tree-sha1 = "2ec0a666381ab2cf04f5b8c42f965256a44b277e"
 uuid = "873c0968-716b-5aa7-bb8d-d1e2e2aeff2d"
-version = "4.2.0+0"
+version = "4.2.0+1"
 
 [[deps.ROCmOpenCLRuntime_jll]]
 deps = ["Artifacts", "JLLWrappers", "Libdl", "Libglvnd_jll", "Pkg", "ROCmCompilerSupport_jll", "ROCmDeviceLibs_jll", "Xorg_libX11_jll", "Xorg_xorgproto_jll", "hsa_rocr_jll", "hsakmt_roct_jll"]
