+++ src/AMDGPU.jl
@@ -18,6 +18,9 @@ export ROCDevice, ROCQueue, ROCExecutable, ROCKernel, ROCSignal
 export has_rocm_gpu
 
 export ROCArray, ROCVector, ROCMatrix, ROCVecOrMat
+export DenseROCArray, DenseROCVector, DenseROCMatrix, DenseROCVecOrMat,
+       StridedROCArray, StridedROCVector, StridedROCMatrix, StridedROCVecOrMat,
+       AnyROCArray, AnyROCVector, AnyROCMatrix, AnyROCVecOrMat
 export roc
 
 # Load HSA Runtime
@@ -69,6 +72,7 @@ module Runtime
     include("safe-load.jl")
 end # module Runtime
 import .Runtime: Mem
+import .Runtime: Adaptor
 
 const ci_cache = GPUCompiler.CodeCache()
 Base.Experimental.@MethodTable(method_table)
@@ -95,13 +99,13 @@ module Device
     include(joinpath("device", "exceptions.jl"))
 end
 import .Device: malloc, signal_exception, report_exception, report_oom, report_exception_frame
-import .Device: ROCDeviceArray, AS, HostCall, hostcall!
+import .Device: ROCDeviceArray, ROCDeviceVector, ROCDeviceMatrix, ROCBoundsError, AS, HostCall, hostcall!
 import .Device: workitemIdx, workgroupIdx, workgroupDim, gridItemDim, gridGroupDim
 import .Device: threadIdx, blockIdx, blockDim
 import .Device: sync_workgroup
 import .Device: @rocprint, @rocprintln, @rocprintf
 
-export ROCDeviceArray
+export ROCDeviceArray, ROCDeviceVector, ROCDeviceMatrix, ROCBoundsError
 export @rocprint, @rocprintln, @rocprintf
 export workitemIdx, workgroupIdx, workgroupDim, gridItemDim, gridGroupDim
 export sync_workgroup
@@ -164,7 +168,7 @@ end
 # Load ROCm external libraries
 if functional(:hip)
 functional(:rocblas)      && include(joinpath(@__DIR__, "blas", "rocBLAS.jl"))
-#functional(:rocsparse)  && include("sparse/rocSPARSE.jl")
+functional(:rocsparse)  && include(joinpath(@__DIR__, "sparse", "rocSparse.jl"))
 #functional(:rocsolver)   && include("solver/rocSOLVER.jl")
 #functional(:rocalution) && include("solver/rocALUTION.jl")
 if functional(:rocrand)
