+++ src/sparse/management.jl
@@ -1,7 +1,7 @@
 # rocSPARSE functions for managing the library
 
-function rocsparse_create()
-    handle = Ref{rocsparse_create_handle}()
-    @check unsafe_rocsparse_create_handle(handle) rocsparse_status_not_initialized
-    handle[]
-end
+# function rocsparse_create()
+#     handle = Ref{rocsparse_create_handle}()
+#     @check unsafe_rocsparse_create_handle(handle) rocsparse_status_not_initialized
+#     handle[]
+# end
